requirements.txt
beautifulsoup4==4.11.1
branca==0.5.0
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
folium==0.12.1.post1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
lightgbm==3.3.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
